+++ src/nvim/window.h
@@ -8,17 +8,6 @@
 #include "nvim/option_defs.h"  // IWYU pragma: keep
 #include "nvim/types_defs.h"  // IWYU pragma: keep
 
-/// Values for file_name_in_line()
-enum {
-  FNAME_MESS  = 1,   ///< give error message
-  FNAME_EXP   = 2,   ///< expand to path
-  FNAME_HYP   = 4,   ///< check for hypertext link
-  FNAME_INCL  = 8,   ///< apply 'includeexpr'
-  FNAME_REL   = 16,  ///< ".." and "./" are relative to the (current)
-                     ///< file instead of the current directory
-  FNAME_UNESC = 32,  ///< remove backslashes used for escaping
-};
-
 /// arguments for win_split()
 enum {
   WSP_ROOM    = 0x01,   ///< require enough room
